hospital-backendprocess/ai-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
sentence-transformers
requests
python-dotenv